requirements.txt
playwright>=1.40.0
aiohttp>=3.9.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
gspread>=6.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
aiofiles>=23.2.1
structlog>=24.1.0
cryptography>=42.0.0
python-dateutil>=2.8.2
python-docx>=1.1.0
pypdf>=4.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
